[compiler v2] Initial setup of bytecode processing
This extends the compiler driver and testsuite by the basic modules for bytecode generation and processing:

- Adds a generator from the Move Model AST to stackless bytecode. This first version only deals with basic expressions.
- Configures a stackless bytecode processor pipeline. This will allow us to add multiple phases like safety and borrow analysis and optimizations. For now the pipeline is empty.
- Adopts the testsuite to support a test configuration which is configured via the test source directory. This allows us to group tests for testing different aspects of a pipeline.

Some of this may look like overkill but this apporach has been worked well for the stackless bytecode processing of the Move Prover. Subsequent PRs will add first processors to the pipeline and iilustrate the usage.